# Remove cuda from autograd profiler