--- a/torch/csrc/autograd/profiler.cpp
+++ b/torch/csrc/autograd/profiler.cpp
@@ -1,13 +1,19 @@
 #include <torch/csrc/autograd/profiler.h>
 #include <torch/csrc/autograd/function.h>
 
-#ifdef USE_CUDA
-#include <c10/cuda/CUDAGuard.h>
-#endif
-
 #include <sstream>
 
 namespace torch { namespace autograd { namespace profiler {
+
+CUDAStubs default_stubs;
+constexpr CUDAStubs* default_stubs_addr = &default_stubs;
+// constant initialization, so it is guarenteed to be initialized before
+// static initialization calls which may invoke registerCUDAMethods
+static CUDAStubs* cuda_stubs = default_stubs_addr;
+
+TORCH_API void registerCUDAMethods(CUDAStubs* stubs) {
+  cuda_stubs = stubs;
+}
 
 ProfilerState state = ProfilerState::Disabled;
 uint16_t next_thread_id = 0;
@@ -31,12 +37,7 @@ void mark(std::string name, bool include_cuda /* = true */) {
     return;
   }
   if (state == ProfilerState::NVTX) {
-#ifdef USE_CUDA
-    nvtxMarkA(name.c_str());
-#else
-    throw std::logic_error(
-        "mark called with NVTX tracing, but compiled without CUDA");
-#endif
+    cuda_stubs->nvtxMarkA(name.c_str());
   } else {
     getEventList().record(
         EventKind::Mark,
@@ -56,18 +57,13 @@ void pushRangeImpl(T name, const char* msg="", int64_t sequence_nr=-1) {
     return;
   }
   if (state == ProfilerState::NVTX) {
-#ifdef USE_CUDA
     if(sequence_nr >= 0) {
       std::stringstream s;
       s << name << msg << sequence_nr;
-      nvtxRangePushA(s.str().c_str());
+      cuda_stubs->nvtxRangePushA(s.str().c_str());
     } else {
-      nvtxRangePushA(c_str(name));
+      cuda_stubs->nvtxRangePushA(c_str(name));
     }
-#else
-    throw std::logic_error(
-        "pushRange called with NVTX tracing, but compiled without CUDA");
-#endif
   } else {
     getEventList().record(
         EventKind::PushRange,
@@ -86,12 +82,7 @@ void popRange() {
     return;
   }
   if (state == ProfilerState::NVTX) {
-#ifdef USE_CUDA
-    nvtxRangePop();
-#else
-    throw std::logic_error(
-        "popRange called with NVTX tracing, but compiled without CUDA");
-#endif
+    cuda_stubs->nvtxRangePop();
   } else {
     getEventList().record(
         EventKind::PopRange,
@@ -103,7 +94,7 @@ void popRange() {
 
 RecordFunction::RecordFunction(Function* fn) {
   // typeid(*fn).name() would avoid an additional string allocation.
-  // However, typeid(*fn).name() would cause nvtx annotations for all user-defined 
+  // However, typeid(*fn).name() would cause nvtx annotations for all user-defined
   // (Python-side) custom autograd function backward() methods to have the same name,
   // because they route through the same C++ side class.
   // fn->name() ensures that nvtx annotations for custom function backward() methods
@@ -124,48 +115,32 @@ RecordFunction::RecordFunction(const char* name, int64_t current_sequence_nr)
   pushRangeImpl<const char*>(name, ", seq=", current_sequence_nr);
 }
 
-#ifdef USE_CUDA
-static void onEachDevice(std::function<void(int)> op) {
-  at::cuda::OptionalCUDAGuard device_guard;
-  int count;
-  TORCH_CUDA_CHECK(cudaGetDeviceCount(&count));
-  for(int i = 0; i < count; i++) {
-    device_guard.set_index(i);
-    op(i);
-  }
-}
-#endif
-
 void enableProfiler(ProfilerState new_state) {
   AT_ASSERT(new_state != ProfilerState::Disabled);
-#ifndef USE_CUDA
-  if (new_state == ProfilerState::NVTX)
+  if (new_state == ProfilerState::NVTX && !cuda_stubs->enabled())
     throw std::runtime_error("Can't use NVTX profiler - PyTorch was compiled without CUDA");
-#endif
   if (state != ProfilerState::Disabled && new_state != state) {
       throw std::runtime_error("can't change kind of profiling (e.g. NVTX to CPU) while profiler is running");
   }
   state = new_state;
 
-#ifdef USE_CUDA
   if(state == ProfilerState::CUDA) {
     // event recording appears to have some startup overhead, so we need to
     // to generate some dummy events first before recording syncrhonization events
     for(int i = 0; i < 5; i++) {
-      onEachDevice([](int d) {
+      cuda_stubs->onEachDevice([](int d) {
           mark("__cuda_startup");
-          cudaDeviceSynchronize();
+          cuda_stubs->synchronize();
       });
     }
 
     // cuda events must be on the same device, so we need a start event recorded
     // for each gpu. we then use this event to synchronize time on the GPU
     // with the CPU clock.
-    onEachDevice([](int d) {
+    cuda_stubs->onEachDevice([](int d) {
         mark("__cuda_start_event");
     });
   }
-#endif
   mark("__start_profile", false);
 }
 
@@ -196,5 +171,25 @@ thread_event_lists disableProfiler() {
     return result;
   }
 }
+
+void Event::record(bool record_cuda) {
+  if (record_cuda) {
+    cuda_stubs->record(&device_, &event, &cpu_ns_);
+    return;
+  }
+  cpu_ns_ = getTime();
+}
+
+double Event::cuda_elapsed_us(const Event & e) {
+  if(!e.has_cuda() || !has_cuda()) {
+    throw std::logic_error("Events were not recorded for CUDA");
+  }
+  if(e.device() != device()) {
+    throw std::logic_error("Events are not on the same device");
+  }
+  return cuda_stubs->elapsed(event, e.event);
+}
+
+CUDAStubs::~CUDAStubs() = default;
 
 }}}