--- a/torch/csrc/autograd/profiler.h
+++ b/torch/csrc/autograd/profiler.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#ifdef USE_CUDA
-#include <nvToolsExt.h>
-#endif
 #include <thread>
 #include <iostream>
 #include <mutex>
@@ -16,20 +13,53 @@
 #include <tuple>
 #include <ATen/ATen.h>
 #include <torch/csrc/WindowsTorchApiMacro.h>
-#include <torch/csrc/cuda/cuda_check.h>
-#ifdef USE_CUDA
-#include <ATen/cuda/CUDAContext.h>
-#include <cuda_runtime.h>
-#endif
 #ifndef _WIN32
 #include <ctime>
 #endif
+
+typedef struct CUevent_st* CUDAEventStub;
 
 namespace torch { namespace autograd {
 
 struct Function;
 
 namespace profiler {
+
+struct TORCH_API CUDAStubs {
+  virtual void record(int* device, CUDAEventStub* event, int64_t* cpu_ns) {
+    fail();
+  }
+  virtual float elapsed(CUDAEventStub event, CUDAEventStub event2) {
+    fail();
+    return 0.f;
+  }
+  virtual void nvtxMarkA(const char* name) {
+    fail();
+  }
+  virtual void nvtxRangePushA(const char* name) {
+    fail();
+  }
+  virtual void nvtxRangePop() {
+    fail();
+  }
+  virtual bool enabled() {
+    return false;
+  }
+  virtual void onEachDevice(std::function<void(int)> op) {
+    fail();
+  }
+  virtual void synchronize() {
+    fail();
+  }
+  virtual ~CUDAStubs();
+
+private:
+  void fail() {
+    AT_ERROR("CUDA used in profiler but not enabled.");
+  }
+};
+
+TORCH_API void registerCUDAMethods(CUDAStubs* stubs);
 
 constexpr inline size_t ceilToMultiple(size_t a, size_t b) {
   return ((a + b - 1) / b) * b;
@@ -64,7 +94,7 @@ enum class EventKind : uint16_t {
   PopRange
 };
 
-struct Event final {
+struct TORCH_API Event final {
   Event(EventKind kind, std::string name, uint16_t thread_id, bool record_cuda)
   : owned_name_(new std::string(std::move(name)))
   , name_ptr_(owned_name_->c_str())
@@ -75,19 +105,7 @@ struct Event final {
   , kind_(kind)
   , thread_id_(thread_id) { record(record_cuda); }
 
-  void record(bool record_cuda) {
-#ifdef USE_CUDA
-    if (record_cuda) {
-      TORCH_CUDA_CHECK(cudaGetDevice(&device_));
-      TORCH_CUDA_CHECK(cudaEventCreate(&event));
-      auto stream = at::cuda::getCurrentCUDAStream();
-      cpu_ns_ = getTime();
-      TORCH_CUDA_CHECK(cudaEventRecord(event, stream));
-      return;
-    }
-#endif
-    cpu_ns_ = getTime();
-  }
+  void record(bool record_cuda);
   std::string kind() const {
     switch(kind_) {
       case EventKind::Mark: return "mark";
@@ -105,29 +123,9 @@ struct Event final {
   double cpu_elapsed_us(const Event & e) {
     return (e.cpu_ns_ - cpu_ns_)/(1000.0);
   }
-  double cuda_elapsed_us(const Event & e) {
-#ifdef USE_CUDA
-    if(!e.has_cuda() || !has_cuda()) {
-      throw std::logic_error("Events were not recorded for CUDA");
-    }
-    if(e.device() != device()) {
-      throw std::logic_error("Events are not on the same device");
-    }
-    TORCH_CUDA_CHECK(cudaEventSynchronize(event));
-    TORCH_CUDA_CHECK(cudaEventSynchronize(e.event));
-    float ms;
-    TORCH_CUDA_CHECK(cudaEventElapsedTime(&ms, event, e.event));
-    return ms*1000.0;
-#else
-    throw std::logic_error("CUDA not enabled");
-#endif
-  }
+  double cuda_elapsed_us(const Event & e);
   bool has_cuda() const {
-#ifdef USE_CUDA
     return event != nullptr;
-#else
-    return false;
-#endif
   }
   int device() const {
     return device_;
@@ -142,9 +140,7 @@ private:
   EventKind kind_;
   uint16_t thread_id_;
   int device_ = -1;
-#ifdef USE_CUDA
-  cudaEvent_t event = nullptr;
-#endif
+  struct CUevent_st* event = nullptr;
 };
 
 // a linked-list of fixed sized vectors, to avoid