Part of separate cuda compilation for libtorch. Stacked PR, read only the last commit.