@zdevito 2019-01-12
This commit adds stubs in the profiler event class. I looked at word-language-model timings to see if this added any overhead and it does not appear to make a difference:

These were recorded using `emit_nvtx/load_nvtx()` so it is going down the cuda path. Note that many aspects of the profiler have bit rotted. seq=... annotations have broken aggregate reporting and CUDA time doesn't appear to get recorded anymore, but these are orthogonal issues to this PR.
 
Before the change:
```
------------------------------------------------  ---------------  ---------------  ---------------  ---------------  ---------------
Name                                                     CPU time        CUDA time            Calls        CPU total       CUDA total
------------------------------------------------  ---------------  ---------------  ---------------  ---------------  ---------------
_masked_scale, seq=0                                  33234.667us          0.000us                3      99704.000us          0.000us
UnbindBackward, stashed seq=916                       95610.500us          0.000us                2     191221.000us          0.000us
detach_                                                 615.636us          0.000us               11       6772.000us          0.000us
cudnn_is_acceptable, seq=738                           1571.000us          0.000us                1       1571.000us          0.000us
embedding_backward, seq=0                             68743.000us          0.000us                1      68743.000us          0.000us
_thnn_fused_lstm_cell_backward, seq=0                 48539.229us          0.000us               70    3397746.000us          0.000us
zero_, seq=736                                        11981.545us          0.000us               11     131797.000us          0.000us
lstm, seq=738                                       7832276.000us          0.000us                1    7832276.000us          0.000us
reshape, seq=0                                         6629.333us          0.000us                3      19888.000us          0.000us
sum, seq=0                                            34375.000us          0.000us                1      34375.000us          0.000us
FusedDropoutBackward, stashed seq=1095                38738.667us          0.000us                3     116216.000us          0.000us
add, seq=0                                            22094.162us          0.000us              340    7512015.000us          0.000us
AddmmBackward, stashed seq=1098                       85462.000us          0.000us                1      85462.000us          0.000us
StackBackward, stashed seq=1092                       44016.000us          0.000us                2      88032.000us          0.000us
LogSoftmaxBackward, stashed seq=1101                  42542.000us          0.000us                1      42542.000us          0.000us
log_softmax, seq=1101                                 32285.000us          0.000us                1      32285.000us          0.000us
_fused_dropout, seq=737                               33428.000us          0.000us                3     100284.000us          0.000us
nll_loss_backward, seq=0                              66137.000us          0.000us                1      66137.000us          0.000us
embedding_dense_backward, seq=0                       64671.000us          0.000us                1      64671.000us          0.000us
torch::autograd::GraphRoot, stashed seq=1103          11791.000us          0.000us                1      11791.000us          0.000us
dropout, seq=737                                      38698.667us          0.000us                3     116096.000us          0.000us
_thnn_fused_lstm_cell, seq=743                        31566.129us          0.000us               70    2209629.000us          0.000us
nll_loss, seq=1102                                    46421.000us          0.000us                1      46421.000us          0.000us
detach                                                 7547.000us          0.000us                2      15094.000us          0.000us
stack, seq=914                                        75782.167us          0.000us                6     454693.000us          0.000us
TBackward, stashed seq=1097                            5720.879us          0.000us              141     806644.000us          0.000us
unbind, seq=738                                       48210.500us          0.000us                6     289263.000us          0.000us
MmBackward, stashed seq=1090                          62223.571us          0.000us              140    8711300.000us          0.000us
_log_softmax, seq=1101                                27628.000us          0.000us                1      27628.000us          0.000us
EmbeddingBackward, stashed seq=736                    73565.000us          0.000us                1      73565.000us          0.000us
nll_loss_forward, seq=1102                            41677.000us          0.000us                1      41677.000us          0.000us
_log_softmax_backward_data, seq=0                     35071.000us          0.000us                1      35071.000us          0.000us
mm, seq=740                                           20673.312us          0.000us              420    8682791.000us          0.000us
torch::autograd::AccumulateGrad                       16964.091us          0.000us               11     186605.000us          0.000us
NllLossBackward, stashed seq=1102                     82045.000us          0.000us                1      82045.000us          0.000us
addmm, seq=1098                                       50224.000us          0.000us                1      50224.000us          0.000us
ThnnFusedLstmCellBackward, stashed seq=1091           56898.757us          0.000us               70    3982913.000us          0.000us
matmul, seq=740                                       24131.029us          0.000us              140    3378344.000us          0.000us
t, seq=739                                             2946.664us          0.000us              703    2071505.000us          0.000us
embedding, seq=736                                    45484.000us          0.000us                1      45484.000us          0.000us
ViewBackward, stashed seq=1100                         9129.000us          0.000us                3      27387.000us          0.000us
as_strided, seq=0                                      2985.000us          0.000us                3       8955.000us          0.000us
view, seq=1096                                         7921.000us          0.000us                4      31684.000us          0.000us
```

After the change:
```
------------------------------------------------  ---------------  ---------------  ---------------  ---------------  ---------------
Name                                                     CPU time        CUDA time            Calls        CPU total       CUDA total
------------------------------------------------  ---------------  ---------------  ---------------  ---------------  ---------------
_masked_scale, seq=0                                  26060.667us          0.000us                3      78182.000us          0.000us
UnbindBackward, stashed seq=916                       90303.000us          0.000us                2     180606.000us          0.000us
detach_                                                 507.818us          0.000us               11       5586.000us          0.000us
cudnn_is_acceptable, seq=738                           1236.000us          0.000us                1       1236.000us          0.000us
embedding_backward, seq=0                             38512.000us          0.000us                1      38512.000us          0.000us
_thnn_fused_lstm_cell_backward, seq=0                 44847.129us          0.000us               70    3139299.000us          0.000us
zero_, seq=736                                        11333.909us          0.000us               11     124673.000us          0.000us
lstm, seq=738                                       7436058.000us          0.000us                1    7436058.000us          0.000us
reshape, seq=0                                         5739.000us          0.000us                3      17217.000us          0.000us
sum, seq=0                                            22966.000us          0.000us                1      22966.000us          0.000us
FusedDropoutBackward, stashed seq=1095                30890.333us          0.000us                3      92671.000us          0.000us
add, seq=0                                            21421.194us          0.000us              340    7283206.000us          0.000us
AddmmBackward, stashed seq=1098                       76172.000us          0.000us                1      76172.000us          0.000us
StackBackward, stashed seq=1092                       38113.000us          0.000us                2      76226.000us          0.000us
LogSoftmaxBackward, stashed seq=1101                  29377.000us          0.000us                1      29377.000us          0.000us
log_softmax, seq=1101                                 29259.000us          0.000us                1      29259.000us          0.000us
_fused_dropout, seq=737                               32681.000us          0.000us                3      98043.000us          0.000us
nll_loss_backward, seq=0                              61281.000us          0.000us                1      61281.000us          0.000us
embedding_dense_backward, seq=0                       34710.000us          0.000us                1      34710.000us          0.000us
torch::autograd::GraphRoot, stashed seq=1103           7565.000us          0.000us                1       7565.000us          0.000us
dropout, seq=737                                      41693.667us          0.000us                3     125081.000us          0.000us
_thnn_fused_lstm_cell, seq=743                        29896.286us          0.000us               70    2092740.000us          0.000us
nll_loss, seq=1102                                    32627.000us          0.000us                1      32627.000us          0.000us
detach                                                 3532.000us          0.000us                2       7064.000us          0.000us
stack, seq=914                                        70307.667us          0.000us                6     421846.000us          0.000us
TBackward, stashed seq=1097                            5400.638us          0.000us              141     761490.000us          0.000us
unbind, seq=738                                       34491.500us          0.000us                6     206949.000us          0.000us
MmBackward, stashed seq=1090                          57755.214us          0.000us              140    8085730.000us          0.000us
_log_softmax, seq=1101                                24763.000us          0.000us                1      24763.000us          0.000us
EmbeddingBackward, stashed seq=736                    42873.000us          0.000us                1      42873.000us          0.000us
nll_loss_forward, seq=1102                            27284.000us          0.000us                1      27284.000us          0.000us
_log_softmax_backward_data, seq=0                     23662.000us          0.000us                1      23662.000us          0.000us
mm, seq=740                                           19317.148us          0.000us              420    8113202.000us          0.000us
torch::autograd::AccumulateGrad                       16174.273us          0.000us               11     177917.000us          0.000us
NllLossBackward, stashed seq=1102                     74221.000us          0.000us                1      74221.000us          0.000us
addmm, seq=1098                                       45038.000us          0.000us                1      45038.000us          0.000us
ThnnFusedLstmCellBackward, stashed seq=1091           51876.157us          0.000us               70    3631331.000us          0.000us
matmul, seq=740                                       23464.550us          0.000us              140    3285037.000us          0.000us
t, seq=739                                             2681.236us          0.000us              703    1884909.000us          0.000us
embedding, seq=736                                    39603.000us          0.000us                1      39603.000us          0.000us
ViewBackward, stashed seq=1100                         7920.000us          0.000us                3      23760.000us          0.000us
as_strided, seq=0                                      2464.333us          0.000us                3       7393.000us          0.000us
view, seq=1096                                         7485.250us          0.000us                4      29941.000us          0.000us
```